Get set of measure download codes that can be used in get_measure_data().

Get set of measure download codes that can be used in get_measure_data().

Usage

get_measure_download_codes()

Value

data

Examples

# \donttest{
get_measure_download_codes()
#> # A tibble: 11 × 3
#>    datasheet_code  datasheet_description                          datasheet_type
#>    <chr>           <chr>                                          <chr>         
#>  1 myh-adm         A datasheet covering the number of admisions … MEASURE-DATA  
#>  2 myh-los         A datasheet covering the average length of st… MEASURE-DATA  
#>  3 myh-cancer      A datasheet covering cancer surgery waiting t… MEASURE-DATA  
#>  4 myh-cws         A datasheet covering financial performance of… MEASURE-DATA  
#>  5 myh-ed          A datasheet covering emergency department wai… MEASURE-DATA  
#>  6 myh-es          A datasheet covering elective surgery waiting… MEASURE-DATA  
#>  7 myh-hh          A datasheet covering hand hygiene rates for p… MEASURE-DATA  
#>  8 myh-hh-private  A datasheet covering hand hygiene rates for p… MEASURE-DATA  
#>  9 myh-sab         A datasheet covering rates and numbers of hea… MEASURE-DATA  
#> 10 myh-sab-private A datasheet covering rates and numbers of hea… MEASURE-DATA  
#> 11 myh-ssi         A datasheet covering the available services o… MEASURE-DATA  
# }